Old behaviour passes tests?

+++ src/precision/CommonBits.cpp
@@ -44,11 +44,12 @@ int64
 CommonBits::zeroLowerBits(int64 bits, int nBits)
 {
 	if (nBits >= 64 || nBits < 0) return 0;
-	int64 invMask = ((int64)1 << nBits) - 1;
+	int64 invMask = (1 << nBits)-1;
 	int64 mask = ~ invMask;
 	int64 zeroed = bits & mask;
 	return zeroed;
 }
+
 
 /*static public*/
 int
